backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
psycopg2-binary
sqlalchemy
asyncpg
google-api-python-client
pgvector
python-multipart
sentence-transformers
langchain-text-splitters
google-auth
google-auth-oauthlib
python-jose[cryptography]
passlib
bcrypt==3.2.2